@@ -9,13 +9,14 @@
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.9.1",
     "jquery": "^3.3.1",
-    "react": "^16.5.2",
+    "react": "16.8",
     "react-accessible-accordion": "^2.4.5",
     "react-bootstrap": "^1.0.0-beta.5",
-    "react-dom": "^16.5.2",
+    "react-dom": "16.8",
     "react-ga": "^2.5.7",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.5",
+    "react-scripts": "3.0.1",
+    "react-scroll": "^1.7.11",
     "react-test-renderer": "^16.8.3",
     "store": "^2.0.12"
   },
@@ -24,5 +25,17 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
